Require the del action to have an ilchtoken
Fix redeclaration of i and original error.

+++ application/modules/war/controllers/admin/Ajax.php
@@ -35,7 +35,7 @@ class Ajax extends Admin
 
     public function delAction()
     {
-        if (!empty($this->getRequest()->getParam('mapid')) && !empty($this->getRequest()->getParam('id'))) {
+        if ($this->getRequest()->isSecure() && !empty($this->getRequest()->getParam('mapid')) && !empty($this->getRequest()->getParam('id'))) {
             $gameMapper = new GamesMapper();
             $gameMapper->deleteById((int)$this->getRequest()->getParam('mapid'));
